Sparse global order reader: refactor merge algorithm. (#3173) (#3182)
* Sparse global order reader: refactor merge algorithm.

While running tests with real world array, I noticed that the
optimization trying to find the the length of a cell slab once a cell
to be merged was found didn't work. Trying to use a binary search across
the whole tile resulted in more comparisons than a linear search since
the cell slab lengths are never that long. Also, next_cell in
ResultCoords didn't use the bitmap, which caused some issues, so
ResultCoords was split into two classes, ResultCoords and
GlobalOrderResultCoords (which has access to the bitmap). This also
allowed to push some of the logic of the merge down into that class to
simplify graetly the merge function.

---
TYPE: IMPROVEMENT
DESC: Sparse global order reader: refactor merge algorithm.

* Addressing feedback from ypatia.

* Addressing feedback from ypatia, part 2.