Sparse global order reader: read fragments with timestamps.
This change allows the sparse global order reader to read fragments
consolidated with timestamps. To do so, there is a first step, after
tile bitmaps and query condition computations to choose, when there are
cells with the same coordinates and no duplicates, the one with the
greatest timestamp. This will turn off bits in the bitmap for cells with
lesser timestamps. Then, before running the merge operation, fragments
with timestamps tiles are checked to see if the last cell of a tile
coincides with the first cell of the next. If so, the cell is removed
from the bitmap. Again this is done only is duplicates are not allowed.
Also, for the merge function, we now chose the cell to merge by looking
at the timestamps and not only take the cell with the highest fragment
index as this might not guarantee the latest cell anymore. Finally, a
new condition was added to stop the merge if the last loaded cell of a
fragment with timestamps is reached (and there are unloaded tiles for
this fragment) as we cannot guarantee that the cell will not be a
duplicate in the next tile of this fragment.

Also included in the change is some missing plumbing for the timestamps
attribute around fragment metadata and tile offsets.

---
TYPE: IMPROVEMENT
DESC: Sparse global order reader: read fragments with timestamps.