Consolidation w timestamps: cell slab length computations fix.
When a fragment with timestamps gets merged, the maximum length of the
cell slab doesn't check if the last cell of the slab is the last cell
of a consolidated fragment with timestamps when there are more tiles
available to merge. In rare cases, this could cause to merge a cell with
the same coordinate twice.

---
TYPE: IMPROVEMENT
DESC: Consolidation w timestamps: cell slab length computations fix.