Dense reader: fixing query conditions with overlapping domains.
When applying query condition on dense fragments, we need to reset the
result bitmap to 1 for fragment domains that aren't the first one. This
is to prevent the case where a cell didn't match the query condition in
a previous fragment but with the new value, it now does.

---
TYPE: IMPROVEMENT
DESC: Dense reader: fixing query conditions with overlapping domains.