Dense reader: improving parallelization. (#3205)
* Dense reader: improving parallelization.

Before, the dense reader would only parallelize on tiles. This change
implements parallelization within a tile too if there are more cores
than tiles.

Also, it used to load all attributes in memory and then copy to the
users buffers. This was done because CellSlabIter is heavy and we could
run less iterations on it. Unfortunately, when the in memory size of
tiles get very large (greater than 4GB) it causes the copy operation to
run slower because the user's buffers are very far from the actual data.
We now load one attribute at a time and to mitigate the CellSlabIter
issue, a new faster iterator (TileCellSlabIter) was implemented that
iterates on the slabs within a tile only.

---
TYPE: IMPROVEMENT
DESC: Dense reader: improving parallelization.

* Disable move and copy for TileCellSlabIter.

* Changing fragment domain from pair to proper class.

* Using dynamic array for tile subarrays to avoid default initialization.