Sparse global order reader: merge algorithm optimization.
This optimizes the max_cell_slab method to find the max value faster
when using cell-by-cell comparison. First the maximum value is found
using the bitmap, if required.

Then, the upper bound to be used in comparison is found by increasing
the search by a power of 2 value. This will ensure that the algorithm
performs equally well when the returned value is small versus big. Once,
the upper power of two is found, a binary search is conducted between
the upper value and the previous value.

---
TYPE: IMPROVEMENT
DESC: Sparse global order reader: merge algorithm optimization.