Result tile with bitmaps: make attributes private.
This changes the ResultTileWithBitmap class attributes to be private
instead of public and proper accessors are created.

Also cleaning up the usage to get rid of magic numbers and usage of
fields for things they are not meant to be used for.

---
TYPE: IMPROVEMENT
DESC: Result tile with bitmaps: make attributes private.