Updates: adding update value APIs.
This adds an API to set update values for updates. The API is very
simple and allows to create an update value, using a field name, value
and size, with the associated API to free it:
tiledb_update_value_alloc and tiledb_update_value_free.

The value is then set on the query using a new API:
tiledb_query_add_update_value.

---
TYPE: IMPROVEMENT
DESC: Updates: adding update value APIs.