Sparse global order reader: disable reader for reads with qc. (#3342)
The sparse global order reader is applying query condition before it
does deduplication. Unfortunately, if a query condition gets rid of a
cell in a newer fragment but not in the older, deduplication will still
merge the old cell. As the fix is not trivial for 2.10, we are going to
disable the reader when there is a query condition enabled.

---
TYPE: IMPROVEMENT
DESC: Sparse global order reader: disable reader for reads with qc.