Sparse global order reader: consider qc results after deduplication.
This fixes an algorithmic issue in the sparse global order reader where
we need to consider the results of a query condition only after we run
deduplication. The reason is the following: if we have data in a later
fragment that replaces previous data, and the query condition would be
false only for the new replacement data, the cells from the new fragment
need to first replace the cells from the earlier fragment in
deduplication, then the new cells can be ignored in the results because
they don't match the query condition.

The behavior prior to this change would be that the new cells would get
erased in the bitmaps, then deduplication would only see the old cells
and return those in the results.

---
TYPE: IMPROVEMENT
DESC: Sparse global order reader: consider qc results after deduplication.