DAG nodes iteration one (#3328)
This PR is a significantly improved version of 3202, including more fully developed implementations and testing for `Sink` and `Source` ports, as well as an improved finite state machine parameterized by policy.

The state machine is documented in fsm.h but the most significant change from 3202 is that it is now "two-phase."  That is, it has two events for the source and two events for the sink.  The source and the sink each have two states, full and empty.  Transition events are "fill" and "push" for the source, and "pull" and "drain" for the source.  Clients of the source essentially use it as follows (in pseudocode):
```text
  create and inject item into source
  invoke fill event
  invoke push event
```
Clients of the sink essentially use it as follows:
```text
   invoke pull event
  extract item from source
  invoke drain event
```
In 3202, the fill/push events and the full/drain event pairs were essentially combined into single events.  The refactoring into two phases was required because the single-event approach was problematic for the sink.  Using it requires the item to be extracted between the pull and drain events.  The previous approach uses some hacks to enable use of a single event, but splitting use of the ports into two separate events is a much cleaner approach.

The finite state machine itself is parameterized by policy (as a mixin via CRTP).  The primary policy class of interest is the `AsyncStateMachine`, which supports asynchronous data transfer between `Source` and `Sink`, using locks and condition variables to synchronize data access.  

The `Sink` and `Source` ports are now fully able to use a state machine to transfer data.  The `Sink` owns a state machine, which both the `Sink` and the `Source` use (as shown above) to transfer data.  The data items are `std::optional<T>`.  

Most of the code for this PR is in the dag/ports subdirectory.  The test subdirectory contains tests for both the finite state machine (unit_fsm.cc) as well as for the `Source` and `Sink` ports (unit_ports.cc).  Similar testing is done in both cases: a series of basic data-transfer tasks as well as a concurrent asynchronous transfer of a large (ish) sequence of integers from `Source` to `Sink`.

---
TYPE: IMPROVEMENT
DESC: Improved version of 3202