Deletes: refactored readers process deletes. (#3374)
This change allows the refactored readers to process deletes conditions.
For regular fragments, the delete conditions are processed only if the
delete timestamp is greater than the fragment start timestamp. For
fragments consolidated with timestamps, the reader needs to generate a
new condition including the timestamp of the delete and apply this
condition to the fragment.

---
TYPE: IMPROVEMENT
DESC: Deletes: refactored readers process deletes.