Deletes: legacy reader process deletes.
This implements the processing of deletes for the legacy reader. As the
legacy reader processes the query condition on cell slabs, and we don't
know for which cell slabs the is or not timestamps, we unfortunately
cannot optimize as we did with the refactored readers by having a
regular delete condition and timestamped delete condition. As we are not
too concerned with the performance of the reader moving forward, we are
always going to use the delete condition with timestamp. At the lower
level, in query condition, we will evaluate the timestamp condition
either on the full fragment, or do it cell by cell, depending on if the
tile has timestamps or not. This also meant that the apply path of query
condition need to pass the fragment metadata at all levels.

---
TYPE: IMPROVEMENT
DESC: Deletes: legacy reader process deletes.