Implement class for TileTuple.
In preparation for future C.41 work, this implements a full class for
`TileTuple` and moves it from being a simple `std::tuple`. Having
constructors for TileTuple will allow in the future to create the
underlying tile objects with the proper size from the get go.

This also cleans up a few places where the default constructor for `Tile`
was used when clearing memory for a particular attribute.

---
TYPE: IMPROVEMENT
DESC: Implement class for TileTuple.