Sparse refactored readers, mark empty fragments as fully loaded early.
This fixes an issue where the refactored readers would not mark a
fragment as fully loaded when the tile ranges for this fragment are
empty. This prevented queries with a large number of fragments but only
a few fragments in the results to run when they should have, with a low
memory budget.

---
TYPE: IMPROVEMENT
DESC: Sparse refactored readers, mark empty fragments as fully loaded early.