Implementation of `DataBlock`, `DataBlock` allocator, `join` view, and updates to `Source` and `Sink` (#3366)
This PR includes updates (and/or implementations) for sc-18059, sc-18060, and sc-18061.  (There was some trouble with rebasing during this PR, so there appear to be many more commits than there actually were -- there are many apparent duplicates.)

The PR includes the following files (omitting CMakeLists.txt and various trivial files):

data_block
├── data_block.cc
├── data_block.h
├── pool_allocator.cc
├── pool_allocator.h
└── test
    ├── unit_data_block.cc
    ├── unit_data_block.h
    ├── unit_pool_allocator.cc
    └── unit_pool_allocator.h

Implementation of `DataBlock` and `DataBlock` allocator.  The allocator is a simple
pool (free list) allocator of fixed-size chunks of bytes.  A `DataBlock` is a
container that has a `shared_ptr` to a single chunk.


ports
├── fsm.cc
├── fsm.h
├── fsm.md
├── policies.h
├── ports.cc
├── ports.h
├── sink_state_machine.svg
├── source_state_machine.svg
└── test
    ├── helpers.h
    ├── pseudo_nodes.cc
    ├── pseudo_nodes.h
    ├── unit_concurrency.cc
    ├── unit_concurrency.h
    ├── unit_fsm.cc
    ├── unit_fsm.h
    ├── unit_ports.cc
    ├── unit_ports.h
    ├── unit_pseudo_nodes.cc
    └── unit_pseudo_nodes.h

Implementation of the `Source` and `Sink` ports, which are the primary data structures
for transferring data between nodes in the task graph.  The transfer is controlled by
a finite state machine (defined in fsm.h and documented in fsm.md).  Unit tests
include tests for the state maching, for the ports themselves, and for the use of
ports in a very rudimentary task graph setting.  For these latter tests we use "pseudo
nodes", which are also tested.  Also included are unit tests to verify that execution
of pseudo nodes does in fact occur concurrently.


utils
├── range_join.cc
├── range_join.h
├── spinlock.cc
├── spinlock.h
└── test
    ├── unit_range_join.cc
    └── unit_spinlock.cc

We needed a few utilities for implementing the classes above.  To create single
contiguous views of multiple `DataBlock`s we implemented a `join` view, which is
similar to `std::ranges::join_view`.  In anticipation of future lightweight synchronization needs
in scheduling, we also implemented a spin lock (or, really, a spin mutex, which can be
locked with `std::lock_guard`, etc.).  Also included are unit tests for both.

---
TYPE: FEATURE
DESC:  Implementation of `DataBlock`, `DataBlock` allocator, `join` view, and updates to `Source` and `Sink`.