VFS: Adding option to disable batching for `read_tiles`. (#3421)
This adds an option so that VFS does reads directly from storage
without performing any batching. For applications with large tiles and
a very small number of tiles per requests, it eliminates a costly buffer
allocation and memcpy in VFS.

The option is exposed in the vfs.disable_batching setting, which is
experimental and can be removed at any time.

---
TYPE: IMPROVEMENT
DESC: VFS: Adding option to disable batching for `read_tiles`.