Dense reader: mode to return coordinates matching a query condition.
This adds in a mode to return the coordinates that match a query condition in the dense reader using a special mode enabled through an experimental configuration flag. It also enables filling the dense coordinates when a query condition is specified, which wasn't possible before.

Moving forward, we can use the exact same mode and code to also copy the data, but that is a lot of work so that can go in a further release.

---
TYPE: IMPROVEMENT
DESC: Dense reader: mode to return coordinates matching a query condition.