Fragment info: allow to pass in an array directory.
This changes `FragmentInfo` so that it can take in an already loaded array directory. This is useful for the consolidation path where the array is opened without fragments, we get information with about the fragments with the `FragmentInfo` APIs, then load fragments to be consolidated on the opened array. This also changes `open_without_fragments` so that array directory lists everything it will need. That way, we can pass in the `ArrayDirectory` from the opened array. A side effect of this refactor is that we could get rid of the hack to pass in the delete tile locations from the fragment info APIs to the array opened without fragments.

---
TYPE: IMPROVEMENT
DESC: Fragment info: allow to pass in an array directory.