Enable building libcurl with zstd support
This adds support for curl to enable zstd compression in linux/macos builds.
Windows uses cmake for building and the curl cmake does not allow for
setting the library paths for zstd. This means the windows superbuild is
not supported. For now we just enable on linux/macos.

The goal of enable zstd to is to allow this as an option for http compression.