Fix reader serialization for old clients. (#3446) (#3472)
* Fix reader serialization for old clients.

Old clients might submit a query using the legacy reader when the server
would select the refactored ones. This causes au issue because the read
state gets deserialized by the client and used to determine query
completion. The server needs to serialize a read state that the client
will understand, so for now, if the client decides to use the legacy
reader, the server needs to process that query with the legacy reader.

Long term, we will want to serialize a member of the query to determine
completion, and the read state will be serialized as a blob not
deserialized by the client. That way, the server can decide to use
whatever reader it wants.

---
TYPE: IMPROVEMENT
DESC: Fix reader serialization for old clients.

* Adding test coverage.