Memory tracker: using the correct type for setting default budget. (#3509)
For very large VCF queries using the legacy reader, the reader cannot
load the tile offsets because we reach 4GB of memory consumption for the
array data. The legacy reader does not set a budget for array data so
really the default value for the memory tracker should be the max uint64
value, not uint32.

---
TYPE: IMPROVEMENT
DESC: Memory tracker: using the correct type for setting default budget.