Sparse global order reader: incomplete reads when hitting memory limits.
When we cannot load all tiles in the result cell slabs in memory for a
specific attribute, we truncate the result cell slabs so that it fits.
When doing this, we have to revert progress in the indexes so that the
continuation picks up those tiles again.

---
TYPE: BUG
DESC: Sparse global order reader: incomplete reads when hitting memory limits.