Dimension labels: Adding ordered dimension label reader.
This adds the reader for converting dimension label ranges to indexes for ordered dimension labels. The ranges are set on the attribute using the internal `set_attribute_ranges` API. The reader uses the fact that the labels are ordered to select the tiles to read using the min/max values for the label attribute. It can then determine which tile ranges (in the index domain) to load. Once the tiles are loaded in memory, it runs a binary search for the index that is contained by the specified ranges.

---
TYPE: IMPROVEMENT
DESC: Dimension labels: Adding ordered dimension label reader.