Sparse GO reader: issue when no tile progress because user buffers full.
This fixes an issue when the `sparse global order reader` didn't make any tile progress since the last iteration because the user buffers were full. The new iteration will not find new coordinate tiles to read/unfilter so the `tmp_result_tiles` is empty. This caused a segfault in `compute_result_tiles` as it divides a value by the size of the passed in result tile which is 0. The fix is to simply exit the function early when the list of result tiles is empty.

---
TYPE: IMPROVEMENT
DESC: Sparse GO reader: issue when no tile progress because user buffers full.