Tile metadata: fixing for ordered writes. (#3527)
* Tile metadata: fixing for ordered writes.

For ordered writes, it is possible to write incomplete tiles. The tile written to disk will add non written cells but with empty data. Unfortunately, the empty cells were considered by the tile metadata generator. To fix this, the tile metadata generator was modified to allow the processing of only a portion of the cells, then the dense tiler can run the generator only on the cells that are actually written.

---
TYPE: IMPROVEMENT
DESC: Tile metadata: fixing for ordered writes.


Co-authored-by: Julia Dark <julia@tiledb.com>