Dimension labels: Adding ordered dimension label reader. (#3541)
* Dimension labels: Adding ordered dimension label reader.

This adds the reader for converting dimension label ranges to indexes for ordered dimension labels. The ranges are set on the attribute using the internal `set_attribute_ranges` API. The reader uses the fact that the labels are ordered to select the tiles to read using the min/max values for the label attribute. It can then determine which tile ranges (in the index domain) to load. Once the tiles are loaded in memory, it runs a binary search for the index that is contained by the specified ranges.

---
TYPE: IMPROVEMENT
DESC: Dimension labels: Adding ordered dimension label reader.

* Addressing feedback from @jp-dark, part 1.

* Use range::typed_data.

* Remove fixed/var duplication + @jp-dark feedback 2

* Changes after offline chat with @jp-dark.

* vector -> array.

* Addressing feedback from @jp-dark.

* One more round of feedback and faster tests.