Global order writer: allow splitting fragments.
This enables to set a value for the global order writer for the desired
fragment size. The writer will then, when it reaches that limit, finish
writing the current fragment and create a new fragment to continue the
operation. At the end of the write, if there is more than one fragment
in progress, a consolidated commit file will be written so that an
atomic file system operation is done to commit the fragments to the
array.

This change also cleans up the clean up code in the writers when the
operation fails and we need to remove a partially written fragment.

---
TYPE: IMPROVEMENT
DESC: Global order writer: allow splitting fragments.