Consolidation: add fragment max size. (#3607)
* Consolidation: add fragment max size.

This adds an experimental setting to consolidation so that the user can
specify a maximum fragment size. When consolidation reaches that size,
it will create continue the operation in a new fragment. The result will
be fragments that are fully disjoint in MBRs.

---
TYPE: IMPROVEMENT
DESC: Consolidation: add fragment max size.

* Clarifications requested by @ihnorton.