Bitsort filter: preallocate filtered buffer for dim tiles.
Currently, the bitsort filter expands the filtered buffer for dimension tiles at the filter level. Unfortunately, when there are multiple chunks processed in parallel, this expansion is not safe and causes issues. The fix is to preallocate the dimension tiles before the filter processes the tiles.

---
TYPE: IMPROVEMENT
DESC: Bitsort filter: preallocate filtered buffer for dim tiles.