Adding attribute ordering checks to ordered dimension label reader.
This adds attribute ordering checks to the ordered dimension label reader. The checks will later be used in consolidation, so the code has been added to `ReaderBase`. The check assumes that each fragment is already ordered by the attribute independently. The algorithm first validates that the array non-empty domain across the first dimension has no holes, meaning that every cell was written. Then, it checks each fragment's non-empty domains and sees if any bounds can be skipped. They can be skipped for the array min/max, or for bounds that were fully overwritten by new fragments. Also, if a bound is right next to another bound, we will only validate the first. Then, bounds can be validated with the tile mins/maxs. Any bound that is tile aligned or bound that is next to another can be validated using this method, which is not computationally expensive. If a bound is not tile aligned, we unfortunately need to load the tile of the next fragment that intersects this bound to read the next or previous value.

---
TYPE: IMPROVEMENT
DESC: Adding attribute ordering checks to ordered dimension label reader.