Dense consolidation: set correct non-empty domain. (#3635)
When performing dense consolidation, we expanded the non empty domain of the read/write operations to full tiles so that the global order writer can write full tiles. This writer has not been built with the same computations that allow to write partial tiles like the ordered writer. This change still uses the tile aligned reads so that the dense reader provide full tiles, with fill values to the global order writer but it enables the consolidator to set a non tile aligned subarray on the writer, which is only used to set the non empty domain. That way, if users write fragments that are not tile aligned and consolidate, they will not get more data after consolidation with fill values.

---
TYPE: IMPROVEMENT
DESC: Dense consolidation: set correct non-empty domain.