Remove `open_arrays_` from StorageManager (#3649)
This member variable is only used for two actual runtime checks that are
both covered by the fact that the Array instance is the one invoking the
methods where those assertions are made. Given that Array should have a
pretty clear idea of whether its open or not, these checks are no longer
useful and this member variable can be removed.