Adding attribute ordering checks to ordered dimension label reader. (#3643)
* Adding attribute ordering checks to ordered dimension label reader.

This adds attribute ordering checks to the ordered dimension label reader. The checks will later be used in consolidation, so the code has been added to `ReaderBase`. The check assumes that each fragment is already ordered by the attribute independently. The algorithm first validates that the array non-empty domain across the first dimension has no holes, meaning that every cell was written. Then, it checks each fragment's non-empty domains and sees if any bounds can be skipped. They can be skipped for the array min/max, or for bounds that were fully overwritten by new fragments. Also, if a bound is right next to another bound, we will only validate the first. Then, bounds can be validated with the tile mins/maxs. Any bound that is tile aligned or bound that is next to another can be validated using this method, which is not computationally expensive. If a bound is not tile aligned, we unfortunately need to load the tile of the next fragment that intersects this bound to read the next or previous value.

---
TYPE: IMPROVEMENT
DESC: Adding attribute ordering checks to ordered dimension label reader.

* Refactor code to verify attribute order

* Fixes ordering check to be strictly increasing/decreasing.
* Renames `AttributeOrderValidationData`->`AttributeOrderValidator` and
  moves it into its own file.
* Moves lambdas for validation logic into `AttributeOrderValidator`.
* Isolates logic for finding which fragments to check, if fragment bounds
  need validation, and if there are any holes into a single method.
* General clean-up (simplify code blocks, add additional comments).

* Remove ensure_continuous_domain_written.

* Fix exception checks.

* Small nits.

* Update to manylinux2014

* bool -> uint8_t fix.

* Adjusting comments.

* Adding private class for validation data.

* Suggestion from @jp-dark

Co-authored-by: Julia Dark <julia@tiledb.com>

* More equality tests.

Co-authored-by: Julia Dark <julia@tiledb.com>
Co-authored-by: Isaiah Norton <isaiah@tiledb.io>