Fix write by label (#3664)
This PR removes checks on the query type and layout from the Subarray C-API methods and adds missing checks to the query strategy constructors. This is needed to fix a bug when writing to a dense array using a label range. In addition, dead code was removed from `Query`.

---
TYPE: BUG
DESC: Fix bad write by label error

TYPE: BREAKING_BEHAVIOR
DESC: For global order reads, an error is now raised if multiple ranges are set when the query is submitted. Previously, an error was raised immediately upon adding another range during subarray creation.

TYPE: BREAKING_BEHAVIOR
DESC: For dense writes, an error is now raised if multiple ranges are set when the query is submitted. Previously, an error was raised immediately upon adding another range during subarray creation.

TYPE: BREAKING_BEHAVIOR
DESC: For sparse writes, an error is now raised if a subarray is set when the query is submitted. Previously, an error was raised when adding ranges to a subarray for a sparse write.