Sparse global order reader: fix tile cleanup when ending an iteration.
This change fixes the tile cleanup code when an iteration is done. When the user sets ranges, the first tile for a fragment might not have index 0. The tile index in the read state (initialized at 0) for a fragment might not be touched if an iteration doesn't merge any cells from that fragment, so it might stay 0. As the cleanup code was looking for an exact match on the tile index before stopping cleanup, this would clean up every tiles. The cleanup now cleans up tiles with an index smaller than the index in the read state which fixes this case.

---
TYPE: IMPROVEMENT
DESC: Sparse global order reader: fix tile cleanup when ending an iteration.