Consolidation plan: implementation for creating a plan using MBRs. (#3696)
* Consolidation plan: implementation for creating a plan using MBRs.

This is an implementation for the consolidation plan using the MBRs. We first batch all fragments that intersect together. Then we take small fragments and combine them is the union of their MBRs don't intersect any other fragments. Finally, we take fragments that are too large according to the passed in desired size and add them to their own node so they get split up.

---
TYPE: IMPROVEMENT
DESC: Consolidation plan: implementation for creating a plan using MBRs.

* Address feedback from @ypatia.