Don't retry curl errors for REST
If we received a curl error we should not resubmit the original request. There
are some curl errors that can occur after we have received partial results and
we might have already deserialized and copied data into the user buffer.
Retries resubmit the initial request not accounting for any partial state.