Throw error when memory budget leads to splitting to unary range.
This changes the subarray partitioner to throw an error when a single tile cannot fit the memory budget, which will lead to processing a single cell at a time and a very slow query. The change also includes a setting that allows the query to proceed without throwing an error if this condition occurs.

---
TYPE: IMPROVEMENT
DESC: Throw error when memory budget leads to splitting to unary range.