Add set of schedulers, a task class, and nodes supporting segmented computation. (#3683)
This PR contains a set of schedulers, a task class, and nodes supporting segmented computation.

### Schedulers

This PR includes implementations (and refinements) for three schedulers:
  - Bountiful scheduler.  Invokes `std::async` for each submitted task.  Uses standard library condition variables for synchronization of tasks.  That is, since tasks are identified 1-1 with threads, we just sleep the thread when the task should sleep.
  - Frugal scheduler.  Similar to the bountiful scheduler, but uses a fixed-size thread pool for thread resources.  Likely prone to starvation deadlock because of the use of `std::condition_variable` for descheduling tasks.
  - Throw-Catch scheduler.  A scheduler with a fixed-size thread pool that manages its own tasks, in a fashion similar to how a typical operating system manages its processes.  The state transitions for this scheduler are managed by the state machine in `task_state_machine.h`.  Note that some design decisions for the task state machine and the port state machine  had to be made for scheduling based on the fact that the throw-catch scheduler action functions don't return (since they throw).  These can be revisited as newer schedulers are brought on line.

The basic scheduler interface consists of two functions:
- `submit(Node&)` which adds the task graph node to the scheduler's list of tasks to be executed.  Tasks are executed lazily, so execution *will not* begin when a node is submitted.
- `sync_wait_all` a blocking call that launches all of the scheduler threads to begin executing tasks. Blocks until all tasks have completed.

The unit tests in unit_scheduler.cc use `TEMPLATE_TEST_CASE` to allow all three schedulers to be used with exactly the same testing code.

### Tasks

The file tasks.h contains an implementation of a basic `Task` class.  The `Task` class wraps up task graph nodes and provides the hooks (and state) so that execution can be managed by the scheduler.  The nodes contained in a Task are executed by calling `Task::resume()`.

### Nodes

The file nodes/segmented_nodes.h contains implementations of task graph nodes that use a "segmented" model of computation.  Execution of the node is accomplished through the `resume()` method. Computation within a node might yield before completion of the function.  When `resume()` is again called, execution begins where the last yield occurred, rather than at the start of `resume()`.  This functionality is provided through a simple implementation of Duff's device.

### Task State Machine and Port State Machine

Operation of the scheduler is accomplished through exchanging of events between the port state machine (in fsm.h) and the task state machine (in task_state_machine.h).  Essentially, every action in one of the state machines invokes an event in the other.  For example, a node may invoke the port `pull` function.  This will inject a `pull` event into the port state machine.  If the `pull` cannot be satisfied, the port state machine will invoke a `wait` action, which will inject a wait event into the task state machine, thus putting the task into a waiting state.  The manner in which these different actions are carried out is determined by port polices and task state machine policies.

### Coda

- More in-depth documentation will be provided in the execution/docs subdirectory in a future PR.
- A performance comparison of the three schedulers using the running sieve example will also be the subject of a future PR.


---
TYPE: FEATURE 
DESC: Add set of schedulers, a task class, and nodes supporting segmented computation.