Implement duration instrument for measuring times in stats. (#3746)
* Implement duration instrument for measuring times in stats.

This change switches stats from using a scoped executor to using a new DurationInstrument class. The class is a simple object that saves the start time at creation and reports duration back to the stats at destruction. It allows the removal of the start_times_ map in the stats object.

---
TYPE: IMPROVEMENT
DESC: Implement duration instrument for measuring times in stats.

* Variable rename.

* Addressing feedback from @ihnorton.

* Addressing feedback from @eric-hughes-tiledb.