Add UTF-8 attribute binary comparison support for QueryConditions (#3766)
Add basic UTF-8 support for QueryConditions on attributes.

This adds a basic binary comparison of UTF-8 data. This means that QueryConditions can work with UTF-8 attribute data while it does not allow for Unicode aware comparisons. A specific description of the algorithm would be:

> UTF-8 string data is comparisons are performed on the underlying Unicode code points without regard for any Unicode aware semantics.

An alternative way of stating that is that we're comparing UTF-8 data with `memcmp` and completely disregard the Unicode properties of the data. Future work will be required to support proper Unicode aware comparisons.

---
TYPE: FEATURE
DESC: Add UTF-8 attribute binary comparison support for QueryConditions

===

* Enable UTF-8 attribute comparisons in queries

Currently this enables a raw comparison of Unicode code points encoded
as UTF-8. Which is to say that this comparison does not respect language
rules or alternative encodings of semantically equivalent strings.

* Add a UTF-8 comparison integration test

* Update query unit tests with UTF-8 tests

* Fix bug in dense reader QC application

We weren't properly accounting for a non-zero src_cell in the buffer
offsets sizing.

* Add query condition test with UTF-8 data

* Extract supported string type test

* Alternate indexing for buffer offsets