Remove deprecated C API function `tiledb_coords` (#3743)
Remove deprecated C API function `tiledb_coords` and a macro `TILEDB_COORDS` that calls it. This PR accomplishes the narrow purpose of removing these symbols from the API as part of the removal process. It does not complete the removal process.

Reintroduce `TILEDB_COORDS` in `tiledb::test` only for test purposes. The symbol `TILEDB_COORDS` continues to be used extensively in test code. This PR moves the symbol into the namespace `tiledb::test` and adjusts a few things to allow the tests to continue to compile. This is a transitional measure only; it's on the road to complete removal, but does not arrive there. 

---
TYPE: C_API
DESC: Remove deprecated C API function `tiledb_coords`

===

* Removed deprecated C API function `tiledb_coords` and a macro `TILEDB_COORDS` that calls it.
Reintroduce `TILEDB_COORDS` in `tiledb::test` only for test purposes.

* clang-format