Split Tile class into different classes for read and write. (#3796)
This is preparation work for the work to remove the extra allocation in VFS. To do this, we'll need tiles on the read path to point inside of a larger memory allocation for the filtered buffers. For writes, it is harder to compute the sizes for the filtered buffers for filters that compress the data so we want to keep it as is for now. Splitting the read tiles and write tiles into separate classes will facilitate the work.

This also changes the old WriterTile class to WriterTileTuple, which matches better the code on the read side which uses TileTuple.

---
TYPE: IMPROVEMENT
DESC: Split Tile class into different classes for read and write.