Fix stack-use-after-free
The point of this test is to make sure thread pools clean up after
themselves when going out of scope. The test as written contained a
stack-use-after-free due to the std::atomic variable being deallocated
while there were still threads running with references to it. This just
moves the std::atomic into a shared_ptr which is then referenced by each
thread pool task.