Array open query v2 support (#3626)
This work is about improving the performance of Queries executed on Cloud by reducing server-side array-open operations. To achieve that we are serializing and returning more information about the array to the client in Array open step, which the client can then send back again along with the query submit request. By transferring state (in this case the Array object) from the server and back, we enable Cloud to skip the expensive operation of accessing storage for opening the array again and again when the client submits queries.

TYPE: IMPROVEMENT
DESC: Query v3: Reduce array open operations on Cloud

Co-authored-by: Isaiah Norton <isaiah@tiledb.io>