Read tiles: get rid of extra VFS allocation.
This moves the code that batches reads in a single allocation up the chain so that the readers can allocate that memory directly and prevent an extra allocation at the VFS layer and extra data copy. The tiles now point into those allocations to access their filtered buffers.

The filtered data is kept inside of a new `FilteredData` class that stores multiple `FilteredDataBlock` objects. The filtered data inside of those blocks is stored in a vector of char at the moment until the data blocks are ready for use.

This also gets rid of the tile cache as it doesn't have performance benefits anymore.

---
TYPE: IMPROVEMENT
DESC: Read tiles: get rid of extra VFS allocation.